// src/containers/AdditionalPOForms/Determination/enum.ts
export enum PO_DETERMINATION_KEY {
  TO = 'to',
  D_DATE = 'dDate',
  FROM = 'from',
  PHONE = 'phone',
  DEPARTMENT = 'department',
  SOLE_SOURCE = 'soleSource',
  EMERGENCY_PROCUREMENT = 'emergencyProcurement',
  REQUEST_FOR_QUOTATIONS_ONE_QUOTE = 'requestForQuotationsOneQuote', // TODO: Tuyen Tran will remove it
  REQUEST_FOR_QUOTATIONS_LOWEST = 'requestForQuotationsLowest', // TODO: Tuyen Tran will remove it
  PRICE_ADJUSTMENT = 'priceAdjustment',
  PRICE_ADJ_PO_NUMBER = 'priceAdjPoNumber',
  PRICE_ADJ_CONTRACT_NUMBER = 'priceAdjContractNumber',
  EXEMPT_PROCUREMENT = 'exemptProcurement',
  EXEMPTION_NUMBER = 'exemptionNumber',
  PROFESSIONAL_SVC_PROCUREMENT = 'professionalSvcProcurement',
  EXTENSION = 'extension',
  BENEFICIAL_REASON = 'beneficialReason',
  OTHER = 'other',
  EXPENDITURE_CONDITION_OTHER = 'expenditureConditionOther',
  REASON_1 = 'reason1',
  REASON_1_REF = 'reason1Ref',
  REASON_2 = 'reason2',
  REASON_2_REF = 'reason2Ref',
  REASON_3 = 'reason3',
  REASON_3_REF = 'reason3Ref',
  REASON_4 = 'reason4',
  REASON_4_REF = 'reason4Ref',
  REASON_5 = 'reason5',
  REASON_5_REF = 'reason5Ref',
  REASON_6 = 'reason6',
  REASON_6_REF = 'reason6Ref',
  REASON_6_PO_NUMBER = 'reason6PoNumber',
  REASON_6_CONTRACT_NUMBER = 'reason6ContractNumber',
  REASON_7 = 'reason7',
  COST_JUSTIFICATION = 'costJustification',
  DEPARTMENT_HEAD = 'departmentHead',
  DEPARTMENT_HEAD_DATE = 'departmentHeadDate',
  APPROVED_DUO = 'approvedDuo',
  APPROVED_DUO_DATE = 'approvedDuoDate',

  REQUEST_FOR_QUOTATIONS_INVITATION_FOR_BID_ONLY_ONE = 'requestForQuotationsInvitationForBidOnlyOne',
  REQUEST_FOR_QUOTATIONS_INVITATION_FOR_BID_LOWEST = 'requestForQuotationsInvitationForBidLowest',
  REQUEST_FOR_PROPOSALS = 'requestForProposals',
}